+++ docs/changelog.md
@@ -59,6 +59,7 @@ Click on the version in the following list to see all changes since the previous
 * docs: add documentation of currently supported syntax
 * feat: support explicit input for labels of the created subtasks
 * feat: support explicit input for components of the created subtasks
+* feat: components can be inherited from parent issue
 
 [Unreleased]: https://github.com/codescape/jira-multiple-subtasks/compare/22.5.0...HEAD
 [22.5.0]: https://github.com/codescape/jira-multiple-subtasks/tree/22.5.0